+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Allow to input login and/or password during build
+- Added `pandoc_path` option
+- Better logging and error catching
+
 # 0.1.0
 
 - Initial release.